updates manifest description. bit of code refactoring
